+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -1,19 +1,23 @@
 import { argent } from './argent/argent';
 import { coinbase } from './coinbase/coinbase';
+import { imToken } from './imToken/imToken';
 import { injected } from './injected/injected';
 import { ledger } from './ledger/ledger';
 import { metaMask } from './metaMask/metaMask';
 import { rainbow } from './rainbow/rainbow';
+import { steak } from './steak/steak';
 import { trust } from './trust/trust';
 import { walletConnect } from './walletConnect/walletConnect';
 
 export const wallet = {
   argent,
   coinbase,
+  imToken,
   injected,
   ledger,
   metaMask,
   rainbow,
+  steak,
   trust,
   walletConnect,
 };