+++ site/components/Header/Header.tsx
@@ -11,9 +11,10 @@ import { Text } from 'components/Text/Text';
 import { vars } from 'css/vars.css';
 import NextLink from 'next/link';
 import React from 'react';
+import pckg from '../../../packages/rainbowkit/package.json';
 import { header, logo, row } from './Header.css';
 
-const RAINBOWKIT_VERSION = '0.1.0';
+const RAINBOWKIT_VERSION = pckg.version;
 
 export function Header({
   darkMode,